dom/browser-element/mochitest/test_browserElement_NoPermission.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=710231
-->
<head>
  <title>Test for Bug 710231</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="browserElementTestHelpers.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=710231">Mozilla Bug 710231</a>

<!--
  Test that an iframe with the |mozbrowser| attribute does not emit
  mozbrowserX events when this page does not have the "browser" permission.
-->

<script type="application/javascript">
"use strict";

SimpleTest.waitForExplicitFinish();

function runTest() {
  var iframe = document.createElement("iframe");
  iframe.setAttribute("mozbrowser", "true");
  document.body.appendChild(iframe);

  iframe.addEventListener("mozbrowserloadstart", function() {
    ok(false, "Should not send mozbrowserloadstart event.");
  });

  iframe.addEventListener("load", function() {
    ok(true, "Got iframe load event.");
    SimpleTest.finish();
  });

  iframe.src = browserElementTestHelpers.emptyPage1;
}

browserElementTestHelpers.setEnabledPref(true);
SpecialPowers.removePermission("browser", document);
addEventListener("testready", runTest);

</script>

</body>
</html>
